Update for Peer Review
This is a great application. Well done! The code covers all the requirements and showcase the coder's mastery in functional programming. One suggestion is to add a filter component so that users can filter films either by names or categories. Another improvement can be done in the future is to make UI prettier by using bootstrap or creative CSS.